+++ src/main/java/subway/domain/station/repository/StationRepository.java
@@ -1,7 +1,7 @@
-package subway.station.repository;
+package subway.domain.station.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import subway.station.domain.Station;
+import subway.domain.station.domain.Station;
 
 public interface StationRepository extends JpaRepository<Station, Long> {
     default Station getById(Long id) {
